+++ CHANGELOG.md
@@ -51,6 +51,7 @@ produce any false negatives.
 
 - Profile header is now less "jarring" when animating
 - Community header is also now less jarring.
+- Simplified onboarding screen
 
 ## [Version 1.0 (53)] - 2023-10-25T09:24:16Z